Move use_semantic_text hook to a seperate folder
